Добавляет разбиение main.js на отдельные js-модули
