+++ src/view/trip-list.js
@@ -1,0 +1,3 @@
+const createTripListTemplate = () => `<ul class="trip-days"></ul>`;
+
+export {createTripListTemplate};
